+++ packages/iov-core/src/signingservercore.ts
@@ -1,5 +1,5 @@
-import { ChainId, PublicIdentity, publicIdentityEquals, TransactionId, UnsignedTransaction } from "@iov/bcp";
-import { UserProfile, WalletId } from "@iov/keycontrol";
+import { ChainId, PublicIdentity, TransactionId, UnsignedTransaction } from "@iov/bcp";
+import { UserProfile } from "@iov/keycontrol";
 
 import { MultiChainSigner } from "./multichainsigner";
 
@@ -89,23 +89,6 @@ export class SigningServerCore {
     reason: string,
     transaction: UnsignedTransaction,
   ): Promise<TransactionId | undefined> {
-    let walletId: WalletId;
-    const wallets = this.profile.wallets.value.filter(wallet => {
-      const firstMatchIndex = this.profile.getIdentities(wallet.id).findIndex(identity => {
-        return publicIdentityEquals(identity, transaction.creator);
-      });
-      return firstMatchIndex !== -1;
-    });
-    switch (wallets.length) {
-      case 0:
-        throw new Error("No wallet found to sign this transation");
-      case 1:
-        walletId = wallets[0].id;
-        break;
-      default:
-        throw new Error("More than one wallets contain the identity to sign this transaction");
-    }
-
     let authorized: boolean;
     try {
       authorized = await this.authorizeSignAndPost(reason, transaction);
@@ -115,7 +98,7 @@ export class SigningServerCore {
     }
 
     if (authorized) {
-      const response = await this.signer.signAndPost(transaction, walletId);
+      const response = await this.signer.signAndPost(transaction);
       return response.transactionId;
     } else {
       return undefined;